major refactor of cli options and xml config, replaced event/action system
parsing the configuration options no longer results in actions getting
created on the scheduler. instead, the master is responsible for
taking the xml config and acting on it to distribute the hosts, plugins,
etc to the appropriate slaves. it uses the cli options as defaults which
get overridden in the case that identical xml configs are supplied.
this has resulted in a simplification of the startup process.

virtual host and process bootstrapping has been separated from the
creation of instances of these modules.

also, topology and dns modules are global and now master owns them.
this will require updating when moving to multiple slaves.

closes #281

+++ src/main/host/shd-cpu.c
@@ -13,8 +13,8 @@
 #include "shadow.h"
 
 struct _CPU {
-    guint frequencyKHz;
-    guint rawFrequencyKHz;
+    guint64 frequencyKHz;
+    guint64 rawFrequencyKHz;
     gdouble frequencyRatio;
     SimulationTime threshold;
     SimulationTime precision;
@@ -23,7 +23,7 @@ struct _CPU {
     MAGIC_DECLARE;
 };
 
-CPU* cpu_new(guint frequencyKHz, gint threshold, gint precision) {
+CPU* cpu_new(guint64 frequencyKHz, guint64 threshold, guint64 precision) {
     utility_assert(frequencyKHz > 0);
     CPU* cpu = g_new0(CPU, 1);
     MAGIC_INIT(cpu);
